backend/requirements.txt
# ProjectMeats Backend Requirements
# Django REST Framework backend for PowerApps migration

# Core Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
dj-database-url==2.1.0

# API Documentation
drf-spectacular==0.27.0
django-filter==25.1

# Development and Testing
django-extensions==3.2.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production (optional for development)
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-decouple==3.8
Pillow==10.1.0  # For ImageField support if needed